# Add Comprehensive Performance Benchmarking for File Parsing