Summary
This PR introduces comprehensive performance benchmarking capabilities for the mojo_csv library, expanding from a single small file benchmark to multi-scale performance testing with datasets of varying sizes.

Changes Made

#### 📊 Enhanced Benchmarking Infrastructure
•  Expanded benchmark suite from single micro-file testing to 5 different file size categories
•  Added comprehensive test datasets:
•  Micro: 3 rows (existing test.csv)
•  Mini: 100 rows (leads-100.csv)
•  Small: 1,000 rows (organizations-1000.csv) 
•  Medium: 100,000 rows (people-100000.csv)
•  Large: 2,000,000 rows (references products-2000000.csv -- not checked in due to no LFS)

#### 🔧 Benchmark Code Improvements
•  Modularized benchmark functions - separate functions for each file size category
•  Errors during bench will exit the run, so we don't get false times
•  Standardized timing methodology with 1,000 iterations per test
•  Improved output formatting with clear categorization and timing results

#### 📈 Updated Performance Documentation
•  Enhanced README.md with comprehensive performance metrics
•  Added benchmark results showing performance scaling across file sizes

Performance Results
The updated benchmarks demonstrate performance scaling:
•  Micro (3 rows): 0.007699ms average
•  Mini (100 rows): 0.241136ms average  
•  Small (1k rows): 1.388513ms average
•  Medium (100k rows): 121.217188ms average
•  Large (2m rows): 3582.876541ms average

Testing
All benchmarks are verified to run successfully with proper error handling and exit codes for reliable performance measurement.

This enhancement provides users with clear performance expectations across different file sizes and establishes a robust foundation for future performance optimizations.